move helper functions to utils.py
