+++ rasa_sdk/knowledge_base/actions.py
@@ -6,182 +6,13 @@ from rasa_sdk.knowledge_base.storage import (
     SCHEMA_KEYS_REPRESENTATION,
     SCHEMA_KEYS_KEY,
 )
+from rasa_sdk.knowledge_base.utils import to_str, SLOT_OBJECT_TYPE, \
+    SLOT_LAST_OBJECT_TYPE, SLOT_ATTRIBUTE, get_attributes_of_object, \
+    reset_attribute_slots, SLOT_MENTION, SLOT_LAST_OBJECT, SLOT_LISTED_OBJECTS, \
+    get_object_name
 
 
-SLOT_MENTION = "mention"
-SLOT_OBJECT_TYPE = "object_type"
-SLOT_ATTRIBUTE = "attribute"
-SLOT_LISTED_OBJECTS = "knowledge_base_listed_objects"
-SLOT_LAST_OBJECT = "knowledge_base_last_object"
-SLOT_LAST_OBJECT_TYPE = "knowledge_base_last_object_type"
-
-
-class ActionKnowledgeBase(Action):
-    """
-    Abstract knowledge base action that can be inherited to create custom actions
-    that are able to interact with a knowledge base.
-    """
-
-    def __init__(self, knowledge_base, use_last_object_mention=True):
-        self.knowledge_base = knowledge_base
-        self.use_last_object_mention = use_last_object_mention
-
-    def _get_object_name(self, tracker):
-        """
-        Get the name of the object the user referred to. Either the NER detected the
-        object and stored its name in the corresponding slot (e.g. "Pasta&Pizza Place"
-        is detected as "restaurant") or the user referred to the object by any kind of
-        mention, such as "first one" or "it".
-
-        Args:
-            tracker: the tracker
-
-        Returns: the name of the actual object (value of key attribute in the
-        knowledge base)
-        """
-        mention = tracker.get_slot(SLOT_MENTION)
-        object_type = tracker.get_slot(SLOT_OBJECT_TYPE)
-
-        # the user referred to the object by a mention, such as "first one"
-        if mention:
-            return self._resolve_mention(tracker)
-
-        # check whether the user referred to the objet by its name
-        object_name = tracker.get_slot(object_type)
-        if object_name:
-            return object_name
-
-        if self.use_last_object_mention:
-            # if no explicit mention was found, we assume the user just refers to the last
-            # object mentioned in the conversation
-            return tracker.get_slot(SLOT_LAST_OBJECT)
-
-        return None
-
-    def _resolve_mention(self, tracker):
-        """
-        Resolve the given mention to the name of the actual object.
-
-        Different kind of mentions exist. We distinguish between ordinal mentions and
-        all others for now.
-        For ordinal mentions we resolve the mention of an object, such as 'the first
-        one', to the actual object name. If multiple objects are listed during the
-        conversation, the objects are stored in the slot 'knowledge_base_listed_objects'
-        as a list. We resolve the mention, such as 'the first one', to the list index
-        and retrieve the actual object.
-        For any other mention, such as 'it' or 'that restaurant', we just assume the
-        user is referring to the last mentioned object in the conversation.
-
-        Args:
-            tracker: the tracker
-
-        Returns: name of the actually object
-        """
-
-        mention = tracker.get_slot(SLOT_MENTION)
-        listed_items = tracker.get_slot(SLOT_LISTED_OBJECTS)
-        last_object = tracker.get_slot(SLOT_LAST_OBJECT)
-        last_object_type = tracker.get_slot(SLOT_LAST_OBJECT_TYPE)
-        current_object_type = tracker.get_slot(SLOT_OBJECT_TYPE)
-
-        if not mention:
-            return None
-
-        if listed_items and mention in self.knowledge_base.ordinal_mention_mapping:
-            idx_function = self.knowledge_base.ordinal_mention_mapping[mention]
-            return idx_function(listed_items)
-
-        # NOTE:
-        # for now we just assume that if the user refers to an object, for
-        # example via "it" or "that restaurant", they are actually referring to the last
-        # object that was detected.
-        if current_object_type == last_object_type:
-            return last_object
-
-    def _to_str(self, object_type, object_dict):
-        """
-        Converts an object to its string representation using the lambda function
-        defined in the schema.
-
-        Args:
-            object_type: the object type
-            object_dict: the object with all its attributes
-
-        Returns: a string that represents the object
-        """
-        representation_func = self.knowledge_base.schema[object_type][
-            SCHEMA_KEYS_REPRESENTATION
-        ]
-        return representation_func(object_dict)
-
-    def _get_attributes_of_object(self, tracker, object_type):
-        """
-        If the user mentioned one or multiple attributes of the provided object_type in
-        an utterance, we extract all attribute values from the tracker and put them
-        in a list. The list is used later on to filter a list of objects.
-
-        For example: The user says 'What Italian restaurants do you know?'.
-        The NER should detect 'Italian' as 'cuisine'.
-        Due to the schema definition we know that 'cuisine' is an attribute of the
-        object type 'restaurant'.
-        Thus, this method returns [{'name': 'cuisine', 'value': 'Italian'}] as
-        list of attributes for the object type 'restaurant'.
-
-        Args:
-            tracker: the tracker
-            object_type: the object type
-
-        Returns: a list of attributes
-        """
-        attributes = []
-
-        if object_type not in self.knowledge_base.schema:
-            return attributes
-
-        for attr in self.knowledge_base.schema[object_type][SCHEMA_KEYS_ATTRIBUTES]:
-            attr_val = tracker.get_slot(attr)
-            if attr_val is not None:
-                attributes.append({"name": attr, "value": attr_val})
-
-        return attributes
-
-    def _reset_attribute_slots(self, tracker, object_type):
-        """
-        Reset all attribute slots of the current object type.
-
-        If the user is saying something like "Show me all restaurants with Italian
-        cuisine.", the NER should detect "restaurant" as "object_type" and "Italian" as
-        "cuisine" object. So, we should filter the restaurant objects in the
-        knowledge base by their cuisine (= Italian). When listing objects, we check
-        what attributes are detected by the NER. We take all attributes that are set,
-        e.g. cuisine = Italian. If we don't reset the attribute slots after the request
-        is done and the next utterance of the user would be, for example, "List all
-        restaurants that have wifi.", we would have two attribute slots set: "wifi" and
-        "cuisine". Thus, we would filter all restaurants for two attributes now:
-        wifi = True and cuisine = Italian. However, the user did not specify any
-        cuisine in the request. To avoid that we reset the attribute slots once the
-        request is done.
-
-        Args:
-            object_type: the object type
-            tracker: the tracker
-
-        Returns: list of slots
-        """
-        slots = []
-
-        if object_type not in self.knowledge_base.schema:
-            return slots
-
-        for attr in self.knowledge_base.schema[object_type][SCHEMA_KEYS_ATTRIBUTES]:
-            attr_val = tracker.get_slot(attr)
-            if attr_val is not None:
-                slots.append(SlotSet(attr, None))
-
-        return slots
-
-
-class ActionQueryKnowledgeBase(ActionKnowledgeBase):
+class ActionQueryKnowledgeBase(Action):
     """
     Action that queries the knowledge base for objects and attributes of an object.
     The action needs to be inherited and the knowledge base needs to be set.
@@ -193,8 +24,9 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
     - add the intent and action to domain file
     """
 
-    def __init__(self, knowldege_base, use_last_object_mention=True):
-        super(ActionQueryKnowledgeBase, self).__init__(knowldege_base, use_last_object_mention)
+    def __init__(self, knowledge_base, use_last_object_mention=True):
+        self.knowledge_base = knowledge_base
+        self.use_last_object_mention = use_last_object_mention
 
     def name(self):
         return "action_query_knowledge_base"
@@ -249,7 +81,7 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
             "Found the following objects of type '{}':".format(object_type)
         )
         for i, obj in enumerate(objects, 1):
-            dispatcher.utter_message("{}: {}".format(i, self._to_str(object_type, obj)))
+            dispatcher.utter_message("{}: {}".format(i, to_str(object_type, obj, self)))
 
     def utter_no_objects_found(self, dispatcher, object_type):
         """
@@ -279,6 +111,9 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
         Returns: list of slots
 
         """
+        print(tracker.current_state())
+        print(tracker.events)
+
         object_type = tracker.get_slot(SLOT_OBJECT_TYPE)
         last_object_type = tracker.get_slot(SLOT_LAST_OBJECT_TYPE)
         attribute = tracker.get_slot(SLOT_ATTRIBUTE)
@@ -313,13 +148,13 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
 
         # get all set attribute slots of the object type to be able to filter the
         # list of objects
-        attributes = self._get_attributes_of_object(tracker, object_type)
+        attributes = get_attributes_of_object(tracker, object_type, self)
         # query the knowledge base
         objects = self.knowledge_base.get_objects(object_type, attributes)
 
         if not objects:
             self.utter_no_objects_found(dispatcher, tracker)
-            return self._reset_attribute_slots(tracker, object_type)
+            return reset_attribute_slots(tracker, object_type, self)
 
         self.utter_objects(dispatcher, object_type, objects)
 
@@ -338,7 +173,7 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
             ),
         ]
 
-        return slots + self._reset_attribute_slots(tracker, object_type)
+        return slots + reset_attribute_slots(tracker, object_type, self)
 
     def _query_attribute(self, dispatcher, tracker):
         """
@@ -354,7 +189,7 @@ class ActionQueryKnowledgeBase(ActionKnowledgeBase):
         object_type = tracker.get_slot(SLOT_OBJECT_TYPE)
         attribute = tracker.get_slot(SLOT_ATTRIBUTE)
 
-        object_name = self._get_object_name(tracker)
+        object_name = get_object_name(tracker, self.knowledge_base.ordinal_mention_mapping, self.use_last_object_mention)
 
         if not object_name or not attribute:
             self.utter_rephrase(dispatcher, tracker)
